remove redundant toFloat call

+++ core/src/main/java/com/pierfrancescosoffritti/androidyoutubeplayer/core/player/views/WebViewYouTubePlayer.kt
@@ -46,7 +46,7 @@ private class YouTubePlayerImpl(private val webView: WebView) : YouTubePlayer {
   }
   override fun seekTo(time: Float) = webView.invoke("seekTo", time)
   override fun setPlaybackRate(playbackRate: PlayerConstants.PlaybackRate) = setPlaybackRate(playbackRate.toFloat())
-  override fun setPlaybackRate(playbackRate: Float) = webView.invoke("setPlaybackRate", playbackRate.toFloat())
+  override fun setPlaybackRate(playbackRate: Float) = webView.invoke("setPlaybackRate", playbackRate)
   override fun toggleFullscreen() = webView.invoke("toggleFullscreen")
   override fun addListener(listener: YouTubePlayerListener) = listeners.add(listener)
   override fun removeListener(listener: YouTubePlayerListener) = listeners.remove(listener)
